fix: align recurring anchor day with JPA integer

+++ payment-service/src/main/resources/db/migration/V5__add_recurring_pix.sql
@@ -9,7 +9,7 @@ CREATE TABLE recurring_pix_schedules (
     status VARCHAR(20) NOT NULL,
     next_occurrence_at TIMESTAMPTZ,
     remaining_occurrences INTEGER NOT NULL,
-    anchor_day SMALLINT NOT NULL,
+    anchor_day INTEGER NOT NULL,
     created_at TIMESTAMPTZ NOT NULL,
     updated_at TIMESTAMPTZ NOT NULL,
